admin: move .gql files into mutations/queries/fragments
