+++ admin/src/fragments/EditArticleInitialFragment.gql
