+++ admin/src/mutations/UpdateDiarySettingsMutation.gql
