+++ admin/src/queries/DiarySettingsQuery.gql
